[ADD] l10n_ar: Do not create sale journal when installing CoA
We do not create any sale journal when we install the Coa since we need
a lot of information in order to configured that the user should provide
when creating the journals.

For demo data we add a context key in order to create the sale journals
when installing the CoA for the demo companies.